LayoutTests/canvas/philip/tests/2d.strokeRect.transform-expected.txt

< [index] >
2d.strokeRect.transform
fillRect is affected by transforms
References: 2d.strokeRect
Actual output:
Expected output:

Passed
